+++ src/types/Settings.ts
@@ -56,6 +56,10 @@ type Settings = {
   };
   mongodb: {
     url: string;
+    indexTTL: {
+      priceTTL: number;
+      datumTTL: number;
+    };
   };
   consensus: {
     retries: number;
@@ -138,6 +142,7 @@ type Settings = {
       verificationInterval: number;
     };
     priceFreshness: number;
+    fetcherInterval: number;
   };
   rpcSelectionStrategy: string;
   feedsCacheRefreshCronRule: string;