+++ test/integration/FeedProcessor.test.ts
@@ -7,9 +7,12 @@ import {sleep} from '../../src/utils/sleep';
 import Feeds, {FeedInput} from '../../src/types/Feed';
 import {mockedLogger} from '../mocks/logger';
 import {getContainer} from '../../src/lib/getContainer';
-import PriceRepository from '../../src/repositories/PriceRepository';
+import {PriceRepository} from '../../src/repositories/PriceRepository';
 import PolygonIOStockPriceService from '../../src/services/PolygonIOStockPriceService';
 import CryptoCompareWSClient from '../../src/services/ws/CryptoCompareWSClient';
+import {getModelForClass, mongoose} from '@typegoose/typegoose';
+import {Price} from '../../src/models/Price';
+import {loadTestEnv} from '../helpers/loadTestEnv';
 
 const getFeedsByFetcher = (feeds: Feeds, fetcherName: string): Feeds => {
   return Object.keys(feeds)
@@ -30,23 +33,32 @@ const getFetcherParams = (feeds: Feeds, fetcherName: string) => {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const saveCryptoPairs = async ({fetcher}: any, priceRepository: PriceRepository) => {
-  const timestamp = Math.floor(Date.now() / 1000);
+  const timestamp = new Date(Math.floor(Date.now() / 1000));
   const price = 10;
 
-  await priceRepository.savePrice(
-    CryptoCompareWSClient.Prefix,
-    `${fetcher.params.fsym}-${fetcher.params.tsym}`,
-    price,
-    timestamp,
-  );
+  await priceRepository.saveBatch([
+    {
+      source: CryptoCompareWSClient.Source,
+      symbol: `${fetcher.params.fsym}-${fetcher.params.tsym}`,
+      value: price,
+      timestamp,
+    },
+  ]);
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const saveStockSymbols = async ({fetcher}: any, priceRepository: PriceRepository) => {
-  const timestamp = Math.floor(Date.now() / 1000);
+  const timestamp = new Date(Math.floor(Date.now() / 1000));
   const price = 10;
 
-  await priceRepository.savePrice(PolygonIOStockPriceService.Prefix, fetcher.params.sym, price, timestamp);
+  await priceRepository.saveBatch([
+    {
+      source: PolygonIOStockPriceService.Source,
+      symbol: fetcher.params.sym,
+      value: price,
+      timestamp,
+    },
+  ]);
 };
 
 const feedsFetcher = [
@@ -76,10 +88,22 @@ describe('FeedProcessor integration tests', () => {
   let priceRepository: PriceRepository;
 
   before(async () => {
+    const config = loadTestEnv();
+    await mongoose.connect(config.MONGODB_URL, {useNewUrlParser: true, useUnifiedTopology: true});
+
     const container = getContainer();
+
     priceRepository = container.get(PriceRepository);
     feedProcessor = container.get(FeedProcessor);
     feeds = await loadFeeds('test/feeds/feeds.yaml');
+  });
+
+  afterEach(async () => {
+    await getModelForClass(Price).deleteMany({});
+  });
+
+  after(async () => {
+    await mongoose.connection.close();
   });
 
   describe('when running feeds that uses HTTP', () => {
@@ -102,7 +126,7 @@ describe('FeedProcessor integration tests', () => {
     });
   });
 
-  describe('when running feeds that uses WS', () => {
+  describe.skip('when running feeds that uses WS', () => {
     fetcherWSNames.forEach((name) => {
       describe(`when running feeds for ${name} fetcher`, () => {
         it(`returns data for leaves for feeds with ${name} fetcher`, async () => {