+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Create FeedAgent
+- Create FeedDataProcessor to get price and data calculated
+
+### Changed
+- WS fetcher saves in mongoDB instead of saving in Redis
+
+### Removed
+- MongoDBPriceAggregator, PriceAggregator and RedisPriceAggregator
 
 ## [7.7.0] - 2023-01-26
 ### Added